mcstas2.wrappers binding generator: give "wrap" method different names for different components, otherwise some compilers could not figure out the right module to load
git-svn-id: svn://danse.us/MCViNE/trunk@285 bd03dbb6-ca95-436e-b6ad-03a95796df24

+++ packages/legacycomponents/mcstas2/mcstas2/wrappers/binding_builder/mm/builder.py
@@ -22,6 +22,7 @@ from temporaryfiles import temporarydir
 from shutil import rmtree
 
 
+
 def build( binding ):
     build_binding( binding )
     export_pythonmodules( binding )
@@ -53,7 +54,9 @@ def export_pythonmodules(binding):
     callmm( tmpdir )
 
     #clean up
-    rmtree( tmpdir )
+    from mcstas2 import DEBUG
+    if not DEBUG:
+        rmtree( tmpdir )
     return
     
 
@@ -84,7 +87,9 @@ def build_binding(binding):
     callmm( tmpdir )
 
     #clean up
-    rmtree( tmpdir )
+    from mcstas2 import DEBUG
+    if not DEBUG:
+        rmtree( tmpdir )
     return
